Add files via upload
This was my recent EDA project which i have done on Bank Transactions Dataset, the new thing is that this project is done with the help of Jupyter Notebook. Mostly i was using vscode for EDA, because i was not to familiar with jupyter, but latter i discover that it was been difficult to do EDA, as on vscode i need to run whole code for single output. So, then I understand how to use Jupyter Notebook, and I made my first project in Jupyter Notebook.